requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aom=3.7.1=h93d8f39_0
appnope=0.1.4=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
audioread=3.0.1=py312hb401068_1
blas=1.0=mkl
bottleneck=1.3.7=py312h32608ca_0
brotli=1.1.0=h0dc2134_1
brotli-bin=1.1.0=h0dc2134_1
brotli-python=1.0.9=py312hcec6c5f_8
bzip2=1.0.8=h6c40b1e_6
ca-certificates=2024.6.2=h8857fd0_0
cairo=1.18.0=h99e66fa_0
certifi=2024.6.2=py312hecd8cb5_0
cffi=1.16.0=py312h38bf5a0_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=py312hecd8cb5_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.1=py312h9230928_0
cycler=0.12.1=pyhd8ed1ab_0
dav1d=1.2.1=h0dc2134_0
debugpy=1.8.1=py312hede676d_0
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
expat=2.6.2=hcec6c5f_0
ffmpeg=6.1.1=gpl_haae13e6_101
filelock=3.13.1=py312hecd8cb5_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_2
fontconfig=2.14.2=h5bb23bf_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.53.0=py312hbd25219_0
freetype=2.12.1=hd8bbffd_0
fribidi=1.0.10=hbcb3906_0
gettext=0.22.5=h5ff76d1_2
gettext-tools=0.22.5=h5ff76d1_2
gmp=6.3.0=h73e2aa4_1
gnutls=3.7.9=h1951705_0
graphite2=1.3.13=h73e2aa4_1003
harfbuzz=8.5.0=h053f038_0
icu=73.2=hf5e326d_0
idna=3.7=py312hecd8cb5_0
importlib-metadata=7.1.0=pyha770c72_0
importlib_metadata=7.1.0=hd8ed1ab_0
intel-openmp=2023.1.0=ha357a0b_43548
ipykernel=6.29.4=pyh57ce528_0
ipython=8.25.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=py312hecd8cb5_0
joblib=1.4.2=pyhd8ed1ab_0
jpeg=9e=h6c40b1e_1
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py312hb401068_0
kiwisolver=1.4.5=py312h49ebfd2_1
krb5=1.21.2=hb884880_0
lame=3.100=hb7f2c08_1003
lazy_loader=0.4=pyhd8ed1ab_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libasprintf=0.22.5=h5ff76d1_2
libasprintf-devel=0.22.5=h5ff76d1_2
libass=0.17.1=h80904bb_1
libblas=3.9.0=1_h508aa58_netlib
libbrotlicommon=1.1.0=h0dc2134_1
libbrotlidec=1.1.0=h0dc2134_1
libbrotlienc=1.1.0=h0dc2134_1
libcblas=3.9.0=5_h0661a58_netlib
libcxx=17.0.6=h88467a6_0
libdeflate=1.17=hb664fd8_1
libedit=3.1.20191231=h0678c8f_2
libexpat=2.6.2=h73e2aa4_0
libffi=3.4.4=hecd8cb5_1
libflac=1.4.3=he965462_0
libgettextpo=0.22.5=h5ff76d1_2
libgettextpo-devel=0.22.5=h5ff76d1_2
libgfortran=5.0.0=13_2_0_h97931a8_3
libgfortran5=13.2.0=h2873a65_3
libglib=2.80.2=h0f68cf7_0
libiconv=1.17=hd75f5a5_2
libidn2=2.3.7=h10d778d_0
libintl=0.22.5=h5ff76d1_2
libintl-devel=0.22.5=h5ff76d1_2
libjpeg-turbo=2.0.0=hca72f7f_0
liblapack=3.9.0=5_h0661a58_netlib
libllvm14=14.0.6=hc8e404f_4
libogg=1.3.4=h35c211d_1
libopenvino=2023.2.0=h93d8f39_0
libopenvino-auto-batch-plugin=2023.2.0=h93d8f39_0
libopenvino-auto-plugin=2023.2.0=h93d8f39_0
libopenvino-hetero-plugin=2023.2.0=h93d8f39_0
libopenvino-intel-cpu-plugin=2023.2.0=h93d8f39_0
libopenvino-ir-frontend=2023.2.0=h93d8f39_0
libopenvino-onnx-frontend=2023.2.0=h93d8f39_0
libopenvino-paddle-frontend=2023.2.0=h93d8f39_0
libopenvino-pytorch-frontend=2023.2.0=h93d8f39_0
libopenvino-tensorflow-frontend=2023.2.0=h93d8f39_0
libopenvino-tensorflow-lite-frontend=2023.2.0=h93d8f39_0
libopus=1.3.1=hc929b4f_1
libpng=1.6.39=h6c40b1e_0
librosa=0.10.2.post1=pyhd8ed1ab_0
libsndfile=1.2.2=h9603cec_1
libsodium=1.0.18=hbcb3906_1
libsqlite=3.46.0=h1b8f9f3_0
libtasn1=4.19.0=hb7f2c08_0
libtiff=4.5.1=hcec6c5f_0
libunistring=0.9.10=h0d85af4_0
libvorbis=1.3.7=h046ec9c_0
libvpx=1.13.1=he965462_0
libwebp-base=1.3.2=h6c40b1e_0
libxml2=2.12.7=h3e169fe_1
libzlib=1.2.13=h87427d6_6
llvm-openmp=18.1.7=h15ab845_0
llvmlite=0.42.0=py312h534208b_1
lz4-c=1.9.4=hcec6c5f_1
markupsafe=2.1.3=py312h6c40b1e_0
matplotlib-base=3.8.4=py312hb6d62fa_2
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mkl=2023.1.0=h8e150cf_43560
mkl-service=2.4.0=py312h6c40b1e_1
mkl_fft=1.3.8=py312h6c40b1e_0
mkl_random=1.2.4=py312ha357a0b_0
mpg123=1.32.6=h73e2aa4_0
mpmath=1.3.0=py312hecd8cb5_0
msgpack-python=1.0.8=py312hc3c9ca0_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nettle=3.9.1=h8e11ae5_0
networkx=3.2.1=py312hecd8cb5_0
numba=0.59.1=py312h04e34b5_0
numexpr=2.8.7=py312hac873b0_0
numpy=1.26.4=py312hac873b0_0
numpy-base=1.26.4=py312h6f81483_0
openh264=2.4.0=h93d8f39_0
openjpeg=2.4.0=h66ea3da_0
openssl=3.3.1=h87427d6_0
p11-kit=0.24.1=h65f8906_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.1=py312he282a81_0
parso=0.8.4=pyhd8ed1ab_0
pcre2=10.43=h0ad2156_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=py312h6c40b1e_0
pip=24.0=py312hecd8cb5_0
pixman=0.43.4=h73e2aa4_0
platformdirs=4.2.2=pyhd8ed1ab_0
pooch=1.8.2=pyhd8ed1ab_0
prompt-toolkit=3.0.47=pyha770c72_0
psutil=5.9.8=py312h41838bb_0
ptyprocess=0.7.0=pyhd3deb0d_0
pugixml=1.14=he965462_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.22=pyhd8ed1ab_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.1.2=pyhd8ed1ab_0
pysocks=1.7.1=py312hecd8cb5_0
pysoundfile=0.12.1=pyhd8ed1ab_0
python=3.12.2=h9f0c242_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.12=4_cp312
pytorch=2.2.2=py3.12_0
pytz=2024.1=py312hecd8cb5_0
pyyaml=6.0.1=py312h6c40b1e_0
pyzmq=26.0.3=py312ha04878a_0
readline=8.2=hca72f7f_0
requests=2.32.2=py312hecd8cb5_0
scikit-learn=1.5.0=py312hc214ba5_1
scipy=1.13.1=py312hb9702fa_0
setuptools=69.5.1=py312hecd8cb5_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.10=h6dc393e_1
soxr=0.1.3=hbf74d83_3
soxr-python=0.3.7=py312h3f2338b_0
sqlite=3.45.3=h6c40b1e_0
stack_data=0.6.2=pyhd8ed1ab_0
svt-av1=1.8.0=h93d8f39_0
sympy=1.12=py312hecd8cb5_0
tbb=2021.8.0=ha357a0b_0
threadpoolctl=3.5.0=pyhc1e730c_0
tk=8.6.14=h4d00af3_0
torchaudio=2.2.2=py312_cpu
torchvision=0.17.2=py312_cpu
tornado=6.4.1=py312hbd25219_0
tqdm=4.66.4=pyhd8ed1ab_0
traitlets=5.14.3=pyhd8ed1ab_0
typing_extensions=4.11.0=py312hecd8cb5_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.1=py312hecd8cb5_0
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.43.0=py312hecd8cb5_0
x264=1!164.3095=h775f41a_2
x265=3.5=hbb4e6a2_3
xz=5.4.6=h6c40b1e_1
yaml=0.2.5=haf1e3a3_0
zeromq=4.3.5=hde137ed_4
zipp=3.19.2=pyhd8ed1ab_0
zlib=1.2.13=h87427d6_6
zstd=1.5.5=hc035e20_2
